EPMRPP-39570. Fix tooltip overflow.
