--- app/src/components/main/tooltips/hoverableTooltip/hoverableTooltip.scss
@@ -1,150 +1,0 @@
-$ARROW_HEIGHT: 6px;
-$ARROW_WIDTH: 16px;
-
-.tooltip-trigger {
-  display: inline-block;
-}
-
-/* stylelint-disable selector-max-attribute */
-.tooltip {
-  &[data-placement*='bottom'] {
-    padding-top: $ARROW_HEIGHT;
-  }
-
-  &[data-placement*='top'] {
-    padding-bottom: $ARROW_HEIGHT;
-  }
-
-  &[data-placement*='left'] {
-    padding-right: $ARROW_HEIGHT;
-  }
-
-  &[data-placement*='right'] {
-    padding-left: $ARROW_HEIGHT;
-  }
-
-  &.no-mobile {
-    @media (max-width: $SCREEN_XS_MAX) {
-      display: none;
-    }
-  }
-
-  &.desktop-only {
-    @media (max-width: $SCREEN_SM_MAX) {
-      display: none;
-    }
-  }
-}
-
-.tooltip-content {
-  padding: 15px;
-  border: 1px solid $COLOR--gray-80;
-  background-color: $COLOR--white-two;
-  box-shadow: 0 1px 3px 0 rgba(0, 0, 0, 0.2);
-  font-family: $FONT-REGULAR;
-  font-size: 13px;
-  color: $COLOR--charcoal-grey;
-  word-wrap: break-word;
-  white-space: pre-wrap;
-}
-
-.tooltip-arrow {
-  position: absolute;
-
-  &::before,
-  &::after {
-    content: '';
-    margin: auto;
-    display: block;
-    width: 0;
-    height: 0;
-    border-style: solid;
-  }
-
-  &::after {
-    position: absolute;
-  }
-
-  &[data-placement*='bottom'] {
-    top: 0;
-    width: $ARROW_WIDTH;
-    height: $ARROW_HEIGHT;
-
-    &::before,
-    &::after {
-      border-width: 0 $ARROW_WIDTH/2 $ARROW_HEIGHT;
-    }
-
-    &::before {
-      border-color: transparent transparent $COLOR--gray-80;
-    }
-
-    &::after {
-      top: 1px;
-      border-color: transparent transparent $COLOR--white-two;
-    }
-  }
-
-  &[data-placement*='top'] {
-    bottom: 0;
-    width: $ARROW_WIDTH;
-    height: $ARROW_HEIGHT;
-
-    &::before,
-    &::after {
-      border-width: $ARROW_HEIGHT $ARROW_WIDTH/2 0;
-    }
-
-    &::before {
-      border-color: $COLOR--gray-80 transparent transparent;
-    }
-
-    &::after {
-      bottom: 1px;
-      border-color: $COLOR--white-two transparent transparent;
-    }
-  }
-
-  &[data-placement*='left'] {
-    right: 0;
-    width: $ARROW_HEIGHT;
-    height: $ARROW_WIDTH;
-
-    &::before,
-    &::after {
-      border-width: $ARROW_WIDTH/2 0 $ARROW_WIDTH/2 $ARROW_HEIGHT;
-    }
-
-    &::before {
-      border-color: transparent transparent transparent $COLOR--gray-80;
-    }
-
-    &::after {
-      top: 0;
-      right: 1px;
-      border-color: transparent transparent transparent $COLOR--white-two;
-    }
-  }
-
-  &[data-placement*='right'] {
-    left: 0;
-    width: $ARROW_HEIGHT;
-    height: $ARROW_WIDTH;
-
-    &::before,
-    &::after {
-      border-width: $ARROW_WIDTH/2 $ARROW_HEIGHT $ARROW_WIDTH/2 0;
-    }
-
-    &::before {
-      border-color: transparent $COLOR--gray-80 transparent transparent;
-    }
-
-    &::after {
-      top: 0;
-      left: 1px;
-      border-color: transparent $COLOR--white-two transparent transparent;
-    }
-  }
-}
-/* stylelint-enable selector-max-attribute */
