+++ app/src/pages/inside/stepPage/stepGrid/defectType/patternAnalyzedLabel/patternAnalyzedTooltip/patternAnalyzedTooltip.scss
@@ -1,11 +1,12 @@
 .pattern-list {
+  max-width: 85vw;
 }
 
 .pattern-list-label {
   font-size: 13px;
   font-family: $FONT-REGULAR;
   color: $COLOR--gray-60;
-  padding-top: 10px;
+  padding-top: 15px;
 }
 
 .pattern-name {
@@ -13,7 +14,6 @@
   color: $COLOR--charcoal-grey;
   font-size: 14px;
   font-weight: 600;
-  max-width: 250px;
   text-overflow: ellipsis;
   overflow: hidden;
   white-space: nowrap;
